+++ src/League/OAuth2/Server/Storage/ClientInterface.php
@@ -50,8 +50,8 @@ interface ClientInterface
      * @param  string     $clientId     The client's ID
      * @param  string     $clientSecret The client's secret (default = "null")
      * @param  string     $redirectUri  The client's redirect URI (default = "null")
-     * @param  string     $grantType    The grant type used in the request
+     * @param  string     $grantType    The grant type used in the request (default = "null")
      * @return bool|array               Returns false if the validation fails, array on success
      */
-    public function getClient($clientId, $clientSecret = null, $redirectUri = null, $grantType);
+    public function getClient($clientId, $clientSecret = null, $redirectUri = null, $grantType = null);
 }