+++ src/League/OAuth2/Server/Storage/ScopeInterface.php
@@ -35,8 +35,8 @@ interface ScopeInterface
      * </code>
      *
      * @param  string     $scope     The scope
-     * @param  string     $clientId  The client ID
-     * @param  string     $grantType The grant type used in the request
+     * @param  string     $clientId  The client ID (default = "null")
+     * @param  string     $grantType The grant type used in the request (default = "null")
      * @return bool|array If the scope doesn't exist return false
      */
     public function getScope($scope, $clientId = null, $grantType = null);